[IMP] Adjust readonly group
