+++ cmo_purchase_group/models/__init__.py
@@ -1,7 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import common
 from . import purchase
-from . import res_partner
-from . import stock
-from . import account_invoice
-from . import product
