+++ cmo_sale_group/models/__init__.py
@@ -1,5 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import common
-from . import res_partner
 from . import sale
-from . import product
